+++ CHANGELOG.md
@@ -7,6 +7,7 @@ NOTES:
 FEATURES:
 
 * **New Data Source:** `pagerduty_service` [GH-141]
+* **New Resource:** `pagerduty_event_rule` [GH-150]
 
 BUG FIXES: